chore: split text control state from kraken state
