+++ kraken/lib/widget.dart
@@ -4,5 +4,6 @@
  */
 
 export 'src/widget/kraken.dart';
+export 'src/widget/text_control.dart';
 export 'src/widget/widget_to_element_adaptor.dart';
 export 'src/widget/element_to_widget_adaptor.dart';
